RequirementMachine: Move some code into a new SimplifySubstitutions.cpp
